Fix a few clumsy printf formats

+++ src/MainNFSD/nfs_worker_thread.c
@@ -468,7 +468,7 @@ int nfs_rpc_get_funcdesc(nfs_request_data_t *preqnfs, nfs_function_desc_t *pfunc
             break;
 
           default:
-            LogDebug(COMPONENT_DISPATCH, "NFS DISPATCHER: NFS Protocol version %d unknown", ptr_req->rq_vers);
+            LogDebug(COMPONENT_DISPATCH, "NFS DISPATCHER: NFS Protocol version %d unknown", (int)ptr_req->rq_vers);
             svcerr_decode(ptr_svc);
             return FALSE;
             break;
@@ -494,7 +494,7 @@ int nfs_rpc_get_funcdesc(nfs_request_data_t *preqnfs, nfs_function_desc_t *pfunc
 
           default:
             LogDebug(COMPONENT_DISPATCH, "NFS DISPATCHER: MOUNT Protocol version %d unknown",
-            ptr_req->rq_vers);
+            (int)ptr_req->rq_vers);
             svcerr_decode(ptr_svc);
             return FALSE;
             break;
@@ -561,7 +561,7 @@ int nfs_rpc_get_funcdesc(nfs_request_data_t *preqnfs, nfs_function_desc_t *pfunc
 #endif                          /* _USE_QUOTA */
   else
     {
-      LogDebug(COMPONENT_DISPATCH, "NFS DISPATCHER: protocol %d is not managed", ptr_req->rq_prog);
+      LogDebug(COMPONENT_DISPATCH, "NFS DISPATCHER: protocol %d is not managed", (int)ptr_req->rq_prog);
       svcerr_decode(ptr_svc);
       return FALSE;
     }							/* switch( ptr_req->rq_prog ) */
